requirements.txt
asgiref==3.8.1
django==4.2.24
sqlparse==0.5.3
typing-extensions==4.13.2
black==24.8.0
djangorestframework==3.16.0